requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
db-dtypes==1.1.1
executing==2.0.1
fastapi==0.104.1
google-api-core==2.13.0
google-auth==2.23.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
icecream==2.1.3
idna==3.4
numpy==1.26.2
packaging==23.2
pandas==2.1.3
proto-plus==1.22.3
protobuf==4.25.0
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
sqlalchemy-bigquery==1.8.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
